refactor: extract file upload feature as a component
