+++ frontend/src/Pages/HomePage.js
@@ -1,59 +1,11 @@
-import React, { useState, useEffect, useMemo } from 'react';
-import {
-  Box,
-  Button,
-  Center,
-  Container,
-  Heading,
-  List,
-  ListIcon,
-  ListItem,
-  Text,
-  VStack,
-} from '@chakra-ui/react';
+import React, { useState, useEffect } from 'react';
 import Layout from '../components/layout';
 import { API_URL } from '../utils/constants';
 import { SalaryTable } from '../components/salaryTable';
-import { useDropzone } from 'react-dropzone';
-import { baseStyle, acceptStyle, rejectStyle } from '../styles/dropzone-styles';
-import { CheckIcon } from '@chakra-ui/icons';
+import { FileUpload } from '../components/fileUpload';
 
 const HomePage = () => {
   const [res, setRes] = useState('');
-  const [filesToSend, setFilesToSend] = useState([]);
-
-  function onDrop(acceptedFiles) {
-    // setFilesToSend(old => [...old, ...acceptedFiles]);
-    setFilesToSend(acceptedFiles);
-  }
-
-  const {
-    acceptedFiles,
-    getRootProps,
-    getInputProps,
-    isDragAccept,
-    isDragReject,
-  } = useDropzone({
-    onDrop,
-    accept: '.csv, application/vnd.ms-excel, text/csv',
-    multiple: false,
-  });
-
-  const files = filesToSend.map(file => (
-    <ListItem key={file.name}>
-      <ListIcon as={CheckIcon} color="green.500" />
-      {file.name}
-    </ListItem>
-  ));
-
-  const style = useMemo(
-    () => ({
-      ...baseStyle,
-      ...(isDragAccept ? acceptStyle : {}),
-      ...(isDragReject ? rejectStyle : {}),
-    }),
-    [isDragReject, isDragAccept]
-  );
 
   useEffect(() => {
     fetch(API_URL + '/users')
@@ -66,51 +18,7 @@ const HomePage = () => {
 
   return (
     <Layout>
-      <Center mb={6}>
-        <Container
-          maxW={'md'}
-          bg={'whiteAlpha.100'}
-          boxShadow={'xl'}
-          rounded={'lg'}
-          p={6}
-          direction={'column'}
-        >
-          <Heading
-            as={'h2'}
-            fontSize={{ base: 'xl', sm: '2xl' }}
-            textAlign={'center'}
-            my={5}
-          >
-            CSV Upload
-          </Heading>
-          <Box {...getRootProps({ style })} m={2}>
-            <input {...getInputProps()} />
-            <Text>Drag and drop / Click here to upload csv file</Text>
-          </Box>
-          <VStack>
-            <List py={2}>{files}</List>
-            <Button
-              onClick={() => {
-                const fileToSend = acceptedFiles[0];
-                const formData = new FormData();
-                formData.append(`file`, fileToSend);
-                fetch(API_URL + '/users/upload', {
-                  method: 'POST',
-                  body: formData,
-                })
-                  .then(response => response.json())
-                  .then(success => console.log(success))
-                  .catch(error => console.log(error));
-
-                setFilesToSend([]);
-              }}
-            >
-              Upload
-            </Button>
-          </VStack>
-        </Container>
-      </Center>
-
+      <FileUpload />
       {res ? <SalaryTable data={res} /> : <></>}
     </Layout>
   );
